+++ tools/auth/auth_test.go
@@ -7,7 +7,7 @@ import (
 )
 
 func TestProvidersCount(t *testing.T) {
-	expected := 28
+	expected := 29
 
 	if total := len(auth.Providers); total != expected {
 		t.Fatalf("Expected %d providers, got %d", expected, total)
@@ -277,5 +277,14 @@ func TestNewProviderByName(t *testing.T) {
 	}
 	if _, ok := p.(*auth.Wakatime); !ok {
 		t.Error("Expected to be instance of *auth.Wakatime")
+	}
+
+	// linear
+	p, err = auth.NewProviderByName(auth.NameLinear)
+	if err != nil {
+		t.Errorf("Expected nil, got error %v", err)
+	}
+	if _, ok := p.(*auth.Linear); !ok {
+		t.Error("Expected to be instance of *auth.Linear")
 	}
 }